Update personal fork.

+++ src/vs/base/node/pfs.ts
@@ -5,58 +5,32 @@
 
 'use strict';
 
-import { Promise, TPromise } from 'vs/base/common/winjs.base';
-import extfs = require('vs/base/node/extfs');
-import paths = require('vs/base/common/paths');
-import { dirname, join } from 'path';
-import { nfcall } from 'vs/base/common/async';
-import fs = require('fs');
-
-export function isRoot(path: string): boolean {
-	return path === dirname(path);
-}
+import { TPromise } from 'vs/base/common/winjs.base';
+import * as extfs from 'vs/base/node/extfs';
+import { join } from 'path';
+import { nfcall, Queue } from 'vs/base/common/async';
+import * as fs from 'fs';
+import * as os from 'os';
+import * as platform from 'vs/base/common/platform';
+import { once } from 'vs/base/common/event';
 
 export function readdir(path: string): TPromise<string[]> {
 	return nfcall(extfs.readdir, path);
 }
 
 export function exists(path: string): TPromise<boolean> {
-	return new Promise(c => fs.exists(path, c));
+	return new TPromise(c => fs.exists(path, c), () => { });
 }
 
 export function chmod(path: string, mode: number): TPromise<boolean> {
 	return nfcall(fs.chmod, path, mode);
 }
 
-export function mkdirp(path: string, mode?: number): TPromise<boolean> {
-	const mkdir = () => nfcall(fs.mkdir, path, mode)
-		.then(null, (err: NodeJS.ErrnoException) => {
-			if (err.code === 'EEXIST') {
-				return nfcall(fs.stat, path)
-					.then((stat: fs.Stats) => stat.isDirectory
-						? null
-						: Promise.wrapError(new Error(`'${ path }' exists and is not a directory.`)));
-			}
-
-			return TPromise.wrapError<boolean>(err);
-		});
-
-	if (isRoot(path)) {
-		return TPromise.as(true);
-	}
-
-	return mkdir().then(null, (err: NodeJS.ErrnoException) => {
-		if (err.code === 'ENOENT') {
-			return mkdirp(dirname(path), mode).then(mkdir);
-		}
-
-		return TPromise.wrapError<boolean>(err);
-	});
-}
+export import mkdirp = extfs.mkdirp;
 
 export function rimraf(path: string): TPromise<void> {
-	return stat(path).then(stat => {
-		if (stat.isDirectory()) {
+	return lstat(path).then(stat => {
+		if (stat.isDirectory() && !stat.isSymbolicLink()) {
 			return readdir(path)
 				.then(children => TPromise.join(children.map(child => rimraf(join(path, child)))))
 				.then(() => rmdir(path));
@@ -65,7 +39,7 @@ export function rimraf(path: string): TPromise<void> {
 		}
 	}, (err: NodeJS.ErrnoException) => {
 		if (err.code === 'ENOENT') {
-			return;
+			return void 0;
 		}
 
 		return TPromise.wrapError<void>(err);
@@ -73,30 +47,30 @@ export function rimraf(path: string): TPromise<void> {
 }
 
 export function realpath(path: string): TPromise<string> {
-	return nfcall(fs.realpath, path, null);
+	return nfcall(extfs.realpath, path);
 }
 
 export function stat(path: string): TPromise<fs.Stats> {
 	return nfcall(fs.stat, path);
 }
 
+export function statLink(path: string): TPromise<{ stat: fs.Stats, isSymbolicLink: boolean }> {
+	return nfcall(extfs.statLink, path);
+}
+
 export function lstat(path: string): TPromise<fs.Stats> {
 	return nfcall(fs.lstat, path);
 }
 
-export function mstat(paths: string[]): TPromise<{ path: string; stats: fs.Stats; }> {
-	return doStatMultiple(paths.slice(0));
-}
-
-export function rename(oldPath: string, newPath: string): Promise {
+export function rename(oldPath: string, newPath: string): TPromise<void> {
 	return nfcall(fs.rename, oldPath, newPath);
 }
 
-export function rmdir(path: string): Promise {
+export function rmdir(path: string): TPromise<void> {
 	return nfcall(fs.rmdir, path);
 }
 
-export function unlink(path: string): Promise {
+export function unlink(path: string): TPromise<void> {
 	return nfcall(fs.unlink, path);
 }
 
@@ -108,19 +82,8 @@ export function readlink(path: string): TPromise<string> {
 	return nfcall<string>(fs.readlink, path);
 }
 
-function doStatMultiple(paths: string[]): TPromise<{ path: string; stats: fs.Stats; }> {
-	let path = paths.shift();
-	return stat(path).then((value) => {
-		return {
-			path: path,
-			stats: value
-		};
-	}, (err) => {
-		if (paths.length === 0) {
-			return err;
-		}
-		return mstat(paths);
-	});
+export function truncate(path: string, len: number): TPromise<void> {
+	return nfcall(fs.truncate, path, len);
 }
 
 export function readFile(path: string): TPromise<Buffer>;
@@ -129,28 +92,53 @@ export function readFile(path: string, encoding?: string): TPromise<Buffer | str
 	return nfcall(fs.readFile, path, encoding);
 }
 
-export function writeFile(path: string, data: string, encoding?: string): Promise;
-export function writeFile(path: string, data: NodeBuffer, encoding?: string): Promise;
-export function writeFile(path: string, data: any, encoding: string = 'utf8'): Promise {
-	return nfcall(fs.writeFile, path, data, encoding);
+// According to node.js docs (https://nodejs.org/docs/v6.5.0/api/fs.html#fs_fs_writefile_file_data_options_callback)
+// it is not safe to call writeFile() on the same path multiple times without waiting for the callback to return.
+// Therefor we use a Queue on the path that is given to us to sequentialize calls to the same path properly.
+const writeFilePathQueue: { [path: string]: Queue<void> } = Object.create(null);
+
+export function writeFile(path: string, data: string, options?: extfs.IWriteFileOptions): TPromise<void>;
+export function writeFile(path: string, data: NodeBuffer, options?: extfs.IWriteFileOptions): TPromise<void>;
+export function writeFile(path: string, data: NodeJS.ReadableStream, options?: extfs.IWriteFileOptions): TPromise<void>;
+export function writeFile(path: string, data: any, options?: extfs.IWriteFileOptions): TPromise<void> {
+	const queueKey = toQueueKey(path);
+
+	return ensureWriteFileQueue(queueKey).queue(() => nfcall(extfs.writeFileAndFlush, path, data, options));
+}
+
+function toQueueKey(path: string): string {
+	let queueKey = path;
+	if (platform.isWindows || platform.isMacintosh) {
+		queueKey = queueKey.toLowerCase(); // accomodate for case insensitive file systems
+	}
+
+	return queueKey;
+}
+
+function ensureWriteFileQueue(queueKey: string): Queue<void> {
+	let writeFileQueue = writeFilePathQueue[queueKey];
+	if (!writeFileQueue) {
+		writeFileQueue = new Queue<void>();
+		writeFilePathQueue[queueKey] = writeFileQueue;
+
+		const onFinish = once(writeFileQueue.onFinished);
+		onFinish(() => {
+			delete writeFilePathQueue[queueKey];
+			writeFileQueue.dispose();
+		});
+	}
+
+	return writeFileQueue;
 }
 
 /**
 * Read a dir and return only subfolders
 */
 export function readDirsInDir(dirPath: string): TPromise<string[]> {
-	return readdir(dirPath).then((children) => {
-		return TPromise.join(
-			children.map((child) => dirExistsWithResult(paths.join(dirPath, child), child))
-		).then((subdirs) => {
-			return removeNull(subdirs);
+	return readdir(dirPath).then(children => {
+		return TPromise.join(children.map(c => dirExists(join(dirPath, c)))).then(exists => {
+			return children.filter((_, i) => exists[i]);
 		});
-	});
-}
-
-function dirExistsWithResult<T>(path: string, successResult: T): TPromise<T> {
-	return dirExists(path).then((exists) => {
-		return exists ? successResult : null;
 	});
 }
 
@@ -169,31 +157,40 @@ export function fileExists(path: string): TPromise<boolean> {
 }
 
 /**
-* Read dir at `path` and return only files matching `pattern`
-*/
-export function readFiles(path: string, pattern: RegExp): TPromise<string[]> {
-	return readdir(path).then((children) => {
-		children = children.filter((child) => {
-			return pattern.test(child);
-		});
-		let fileChildren = children.map((child) => {
-			return fileExistsWithResult(paths.join(path, child), child);
-		});
-		return TPromise.join(fileChildren).then((subdirs) => {
-			return removeNull(subdirs);
-		});
+ * Deletes a path from disk.
+ */
+let _tmpDir: string = null;
+function getTmpDir(): string {
+	if (!_tmpDir) {
+		_tmpDir = os.tmpdir();
+	}
+	return _tmpDir;
+}
+export function del(path: string, tmp = getTmpDir()): TPromise<void> {
+	return nfcall(extfs.del, path, tmp);
+}
+
+export function whenDeleted(path: string): TPromise<void> {
+
+	// Complete when wait marker file is deleted
+	return new TPromise<void>(c => {
+		let running = false;
+		const interval = setInterval(() => {
+			if (!running) {
+				running = true;
+				fs.exists(path, exists => {
+					running = false;
+
+					if (!exists) {
+						clearInterval(interval);
+						c(null);
+					}
+				});
+			}
+		}, 1000);
 	});
 }
 
-export function fileExistsWithResult<T>(path: string, successResult: T): TPromise<T> {
-	return fileExists(path).then((exists) => {
-		return exists ? successResult : null;
-	}, (err) => {
-		return TPromise.wrapError(err);
-	});
-}
-
-
-function removeNull<T>(arr: T[]): T[] {
-	return arr.filter(item => (item !== null));
+export function copy(source: string, target: string): TPromise<void> {
+	return nfcall(extfs.copy, source, target);
 }
